Create store function componente global type
