--- react/typings/global.ts
@@ -1,20 +1,0 @@
-type QuantityDisplayType = 'always' | 'never' | 'not-empty'
-
-type MinicartTotalItemsType =
-  | 'total'
-  | 'distinct'
-  | 'totalAvailable'
-  | 'distinctAvailable'
-
-type SlideDirectionType =
-  | 'horizontal'
-  | 'vertical'
-  | 'rightToLeft'
-  | 'leftToRight'
-
-type MinicartVariationType =
-  | 'popup'
-  | 'drawer'
-  | 'link'
-  | 'popupWithLink'
-  | 'block'
